+++ elide-core/src/test/java/example/Publisher.java
@@ -47,7 +47,7 @@ public class Publisher {
         return books;
     }
 
-    public void setBooks(Set<Book> authors) {
+    public void setBooks(Set<Book> books) {
         this.books = books;
     }
 }