+++ elide-integration-tests/src/test/java/example/ExceptionThrowingBean.java
@@ -45,7 +45,7 @@ public class ExceptionThrowingBean {
 
     @Override
     public boolean equals(Object obj) {
-        if (obj == null || !(obj instanceof LineItem)) {
+        if (obj == null || !(obj instanceof ExceptionThrowingBean)) {
             return false;
         }